+++ crates/poe_tree/src/consts.rs
@@ -1,10 +1,10 @@
 //$ crates/poe_tree/src/consts.rs
-pub const ORBIT_RADII: [f64; 10] = [
+pub const ORBIT_RADII: [f32; 10] = [
     0.0, 82.0, 162.0, 335.0, 493.0, 662.0, 846.0, 251.0, 1080.0, 1322.0,
 ];
-pub const ORBIT_SLOTS: [usize; 10] = [1, 12, 24, 24, 72, 72, 72, 24, 72, 144];
+pub const ORBIT_SLOTS: [u32; 10] = [1, 12, 24, 24, 72, 72, 72, 24, 72, 144];
 
-pub const CHAR_START_NODES: [usize; 6] = [
+pub const CHAR_START_NODES: [u32; 6] = [
     50459, // Ranger 4 o'clock
     47175, // Warrior 8'oclock
     50986, // Mercenary 6'oclock
@@ -13,7 +13,7 @@ pub const CHAR_START_NODES: [usize; 6] = [
     44683, // Monk 2'clock
 ];
 
-pub const LEVEL_ONE_NODES: [usize; 12] = [
+pub const LEVEL_ONE_NODES: [u32; 12] = [
     // monk
     10364, 52980, 
     // ranger
