+++ crates/poe_tree/src/coordinates.rs
@@ -1,6 +1,6 @@
 //$ crates/poe_tree/src/coordinates.rs
 #[derive(Debug, Clone)]
 pub struct Group {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
