+++ crates/poe_tree/src/debug_utils.rs
@@ -5,11 +5,11 @@ pub fn format_bytes(bytes: usize) -> String {
     const GB: usize = 1024 * MB;
 
     if bytes >= GB {
-        format!("{:.2} GB", bytes as f64 / GB as f64)
+        format!("{:.2} GB", bytes as f32 / GB as f32)
     } else if bytes >= MB {
-        format!("{:.2} MB", bytes as f64 / MB as f64)
+        format!("{:.2} MB", bytes as f32 / MB as f32)
     } else if bytes >= KB {
-        format!("{:.2} KB", bytes as f64 / KB as f64)
+        format!("{:.2} KB", bytes as f32 / KB as f32)
     } else {
         format!("{} bytes", bytes)
     }
