+++ crates/poe_tree/src/type_wrappings.rs
@@ -1,4 +1,4 @@
 //$ crates/poe_tree/src/type_wrappings.rs
-pub type GroupId = usize;
-pub type NodeId = usize;
-pub type EdgeId = usize;
+pub type GroupId = u32;
+pub type NodeId = u32;
+pub type EdgeId = u32;
